requirements.txt
pandas==2.2.2
fastapi==0.111.0
uvicorn==0.30.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
spacy==3.8.2
# SpaCy Model (Large Spanish) - Added as direct link for easier installation
es_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.8.0/es_core_news_lg-3.8.0-py3-none-any.whl
pyarrow==16.1.0
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
